+++ app/BestSell.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App;
+
+use Illuminate\Database\Eloquent\Model;
+
+class BestSell extends Model
+{
+    //
+}
