+++ app/Http/Controllers/BestSellController.php
@@ -1,0 +1,1 @@
+<?php
