+++ public/js/main.js
@@ -5,3 +5,13 @@ $("#Glide").glide({
     touchDistance: 2
 
 });
+
+
+// search_ajax
+$('#search_text').autocomplete({
+    source: "/search/autocomplete",
+    minLength:1,
+    select: function(event, ui) {
+        $('#search_text').val(ui.item.value);
+    }
+});
