+++ resources/views/template/template.blade.php
@@ -20,6 +20,11 @@
     <link rel="stylesheet" href="css/glide.core.min.css">
     <link rel="stylesheet" href="css/glide.theme.css">
 
+    {{--j-query-ui--}}
+
+    <link rel="stylesheet" href="js/jquery-ui.min.css"></link>
+    <link rel="stylesheet" href="js/jquery-ui.theme.min.css"></link>
+
     <!-- Main style -->
     <link rel="stylesheet" href="css/main.css">
     <link rel="stylesheet" href="css/search.css">
@@ -27,22 +32,11 @@
     <!-- Jquery UI -->
     <script src="js/jquery.js"></script>
     <script src="js/jquery-ui.js"></script>
-    <link rel="stylesheet" href="js/jquery-ui.min.css"></link>
-    <link rel="stylesheet" href="js/jquery-ui.theme.min.css"></link>
-
 </head>
 
 <body>
 <!--Header-->
 @include('includes.header')
-
-@if(Request::is('/'))
-<!--introduce title-->
-@include('includes.introduce')
-<!--news slide-->
-@include('includes.slide')
-@endif
-
 @yield('content')
 
 </body>
